Moved ScatterWidget, ImageWidget, and HistogramWidget (and associated tests) to glue.viewers
