+++ glue/core/state_path_patches.txt
@@ -10,3 +10,6 @@ glue.clients.layer_artist.ScatterLayerArtist -> glue.viewers.scatter.layer_artis
 glue.clients.layer_artist.ImageLayerArtist -> glue.viewers.image.layer_artist.ImageLayerArtist
 glue.clients.layer_artist.HistogramLayerArtist -> glue.viewers.histogram.layer_artist.HistogramLayerArtist
 glue.clients.ds9norm.DS9Normalize -> glue.viewers.image.ds9norm.DS9Normalize
+glue.qt.widgets.scatter_widget.ScatterWidget -> glue.viewers.scatter.qt.ScatterWidget
+glue.qt.widgets.image_widget.ImageWidget -> glue.viewers.image.qt.ImageWidget
+glue.qt.widgets.histogram_widget.HistogramWidget -> glue.viewers.histogram.qt.HistogramWidget
