+++ glue/plugins/ginga_viewer/tests/test_qt_widget.py
@@ -7,7 +7,7 @@ from glue.tests.helpers import GINGA_INSTALLED
 if not GINGA_INSTALLED:
     pytest.skip()
 
-from glue.qt.widgets.tests.test_image_widget import _TestImageWidgetBase
+from glue.viewers.image.qt.tests.test_viewer_widget import _TestImageWidgetBase
 
 from ..qt_widget import GingaWidget
