+++ glue/plugins/tests/test_plotly.py
@@ -4,7 +4,9 @@ import numpy as np
 
 from glue.core import Data, DataCollection
 from glue.qt.glue_application import GlueApplication
-from glue.qt.widgets import HistogramWidget, ScatterWidget
+
+from glue.viewers.scatter.qt import ScatterWidget
+from glue.viewers.histogram.qt import HistogramWidget
 
 from ..export_plotly import build_plotly_call
