+++ glue/plugins/tools/pv_slicer.py
@@ -3,13 +3,13 @@ from __future__ import absolute_import, division, print_function
 import numpy as np
 
 from glue.qt.mouse_mode import PathMode
-from glue.qt.widgets.image_widget import StandaloneImageWidget
+from glue.viewers.image.qt import StandaloneImageWidget
 from glue.qt.widgets.mpl_widget import defer_draw
 
 
 def setup():
     from glue.config import tool_registry
-    from glue.qt.widgets import ImageWidget
+    from glue.viewers.image.qt import ImageWidget
     tool_registry.add(PVSlicerTool, widget_cls=ImageWidget)
