+++ glue/qt/widgets/__init__.py
@@ -1,12 +1,10 @@
 from __future__ import absolute_import, division, print_function
 
 from .custom_component_widget import CustomComponentWidget
-from .histogram_widget import HistogramWidget
-from .image_widget import ImageWidget
 from .layer_tree_widget import LayerTreeWidget
 from .message_widget import MessageWidget
 from .mpl_widget import MplWidget
-from .scatter_widget import ScatterWidget
 from .table_widget import TableWidget
 
-default_widgets = [ScatterWidget, HistogramWidget, ImageWidget]
+# default_widgets = [ScatterWidget, HistogramWidget, ImageWidget]
+default_widgets = []
