+++ glue/viewers/histogram/qt/options_widget.ui
