+++ glue/viewers/histogram/qt/viewer_widget.py
@@ -1,5 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
+import os
 from functools import partial
 
 from glue.external.qt.QtCore import Qt
@@ -49,7 +50,8 @@ class HistogramWidget(DataViewer):
         self.central_widget = MplWidget()
         self.setCentralWidget(self.central_widget)
         self.option_widget = QtGui.QWidget()
-        self.ui = load_ui('histogramwidget', self.option_widget)
+        self.ui = load_ui('options_widget.ui', self.option_widget,
+                          directory=os.path.dirname(__file__))
         self._tweak_geometry()
         self.client = HistogramClient(self._data,
                                       self.central_widget.canvas.fig,
