+++ glue/viewers/image/qt/__init__.py
@@ -1,0 +1,1 @@
+from .viewer_widget import *
