+++ glue/viewers/image/qt/options_widget.ui
